Support 'Duplicate' Products + Perf Improvement
Summary:
When you'd duplicate a product, the FBid would be copied over as well,
which would lead to disastrous situations where many products were tied to one FB product.

The solution is simple : clear the meta FBID and FBProductGroup ID from the product on duplication. WooCommerce provides a simple filter to do this.

Additionally, I moved around 2 functions which I think will benefit perf.
These functions were loading code in a non-admin context, but they are only used in an admin context.
I force resynced to verify that there was no issue with this portion of the code.

Reviewed By: mengyingdu

Differential Revision: D6540967

fbshipit-source-id: d7d1189