WPML Compatibility : Products in non-default Language set to Staging
Summary:
Before this diff, we synced all products in all languages to FB and all were visible in the shop.
Now, all the products which are not in the WPML default language will be marked as 'staging' and not visibile to customers.

We still sync those products however, because having these products will allow ViewContent and AddToCart pixel fires to work properly with multi-language products. If we were to not sync these products, those pixel fires would fail to match, and there would be match rate problems, less than ideal custom audiences, etc.

Reviewed By: mengyingdu

Differential Revision: D6558808

fbshipit-source-id: c280a92