Fix Gallery Images of Variable Product for Feed
Summary:
D5909151 has a fix for this issue since for variable products, gallery images is the parent level concept. In Feed, this feature is broken.  Cache gallery images from parent and only one time for a group.

Also, in feed uploading approach, we limit the # of additional images to 5. This is for the concern of memory and performance.

Reviewed By: dmitridr

Differential Revision: D7296743

fbshipit-source-id: ae083c3