Potential Fix for Compatibility with Enhanced ECommerce Plugin
Summary: I suspect the issue is because our `ViewContent` and similar events get injected in the middle of the Google Analytics events, which causes some interference. I wasn't able to get a repro of the original problem, but this change should prevent the interference from all plugins which have this behavior.

Reviewed By: mengyingdu

Differential Revision: D7733900

fbshipit-source-id: bcaab64