Sync composite product with calculating the price based on the price of sub-items
Summary:
When the regular price does not exist, Woo Product won't be synced to FB. However, this is the usual case for Composite Product.
So we need to check the product type of a product, then call a WC_Product_Composite's get_price method.

This diff simply checks the product type, and calls the corresponding method in WC_Product_Composite class

Reviewed By: dmitridr

Differential Revision: D13034929

fbshipit-source-id: 50faf9f5