Manually revert D13037861
Summary: We found that D13037861 introduced a bug that because existing products don't have this sync_status flag checked, they are not uploaded. Since we introduced a couple other conditions where a product won't be synced. I'm going to revert this diff and test all the combinations before submitting a new diff

Reviewed By: mengyingdu

Differential Revision: D13217619

fbshipit-source-id: 4c2849e7