Initial Support for Advanced Bulk Edit
Summary:
This diff adds initial support for
`Advanced Bulk Edit`. True support would probably happen if the REST api issue is solved, since bulk edit calls the API hooks when updating products.

In any case, what this diff adds is a notice at the top of every admin page saying that the products are out of sync with FB. This is the same thing we do for WP_All_Import. See the video in the test plan.

You have to refresh to see the admin notice though.

I also added a param to the URL for our configuration page to allow removal of the resync sticky. I think it makes sense to remove the sticky once you are already on the page, rather than having the user do an extra click to remove it.

Also the sticky will automatically be removed if product sync occurs through any other way (for example, cron scheduling).

Reviewed By: mengyingdu

Differential Revision: D13071496

fbshipit-source-id: ccba6b76