Remove Code Incompatible with Wordpress Guidelines
Summary:
I removed some code in the wooupdater file, which is not really necessary for the plugin and is against wordpress.org guidelines.
As far as I understand this code tries to install a WooThemes plugin which does auto-update for our plugin as well as others. Most likely, if users have installed from woocommerce.com they already have WooThemes anyway so the code is not needed. The plugin remains compatible with woothemes, but will not autodownload the plugin or surface a notice anymore.

I also added a directive which our push script can use to remove the GitHub autoupdater
before pushing to wordpress.org. The code will remain in the repository and in the woocommerce.com version of the plugin, but will not be included in the wordpress.org version.

Reviewed By: mengyingdu

Differential Revision: D13265126

fbshipit-source-id: 0ad11d32