Added filter for product ID on purchase event (#759)
Summary:
This filter would provide compatibility with plugins which make use of the WC checkout process and order management, but are selling something else, not WC Products directly.

For example, we've encountered this issue with the plugin LearnPress in conjunction with the extension LearnPress - WooCommerce Payments which enables the selling of access to courses (LearnPress) through the WooCommerce checkout process. In this case, the product ID is not stored inside $item['product_id'] but inside $item['course_id']. Since the value of $item['product_id'] is 0 in this case, the code breaks the checkout confirmation page.
Pull Request resolved: https://github.com/facebookincubator/facebook-for-woocommerce/pull/759

Reviewed By: mengyingdu

Differential Revision: D13009976

Pulled By: dmitridr

fbshipit-source-id: 39ec8966