Remove get_product_catalog_id_box()
Handlers for ajax_sync_all_fb_products_using_feed and ajax_sync_all_fb_products
Ajax actions are responsible for verifying that an access token and
catalog ID are correctly set before attempting to sync.

I think there is no need to check for those values in the frontend,
as it was done in sync_all_products(), but we should probably display errors
when actions cannot be performed.