Display successful upload status for feed pull
Summary:
For FBE v1.0, the feed synchronization is migrating from push to pull. This means that the product feed is not being uploaded via API, but rather the global url of the feed is being provided. As a result, there is no need to provide an upload status for the pull mechanism.

This change is somewhat temporary, as FBE v2.0 will revert back to including a push mechanism, which will require the inclusion of an upload status shown to the user. For this diff, a successful upload status is hardcoded for v1.0 as a quick fix. The reversion of this change will be included in the set of FBE v2.0 specific changes coming soon.

Reviewed By: zanyaziz

Differential Revision: D19981966

fbshipit-source-id: e70d51fb