Addition of classes to send server side events async
Summary:
A new class is added to send server events asynchronously on the wp_footer action.
The classes WC_Facebookcommerce_ServerEventSender and WC_Facebookcommerce_Pixel were modified to send the events on the wp_footer action, because we found that we cannot separately send two events at the same time(this happens when you add an item to your cart).

Reviewed By: vinothsa

Differential Revision: D21700113

fbshipit-source-id: 9ce868af