Use relative paths rather than absolute root paths
This is to solve the following warning:
PHP Warning: is_readable(): open_basedir restriction in effect. File(/includes/fbwpml.php) is not within the allowed path(s) [...]

The is_readable call is using a system-root url, i.e. it's looking for `/includes/fbwpml.php` rather than `/path-to-wp/wp-content/plugins./facebook-for-woocommerce/includes/fbwpml.php` which causes a warning similar to the above for any system with open_basedir restrictions that don't allow this.

Also, the call to is_readable will always return false with these paths.

Removes the negation on is_readable also, to take the conditional from `IF class does not exist AND file _is not_ readable` to `IF class does not exist AND file _is_ readable`