Make code compatible with 2.0.0 action scheduler job framework.
We move processing of the products from handle_batch to process_items.
This allows us bypassing processing of items one by one. We can now pass all of the items we want to a function that will do a whole bunch - this function is part of the old processing chain. This is why we are not implementing it here.